Rename package akima2 to makima
Updated package name references from akima2 to makima.
This aligns the package naming with the intended project structure.
