+++ graph/src/main/java/org/akazukin/graph/algorithm/algorithms/makima/ModifiedAkimaCurveGraphBuilder.java
@@ -1,4 +1,4 @@
-package org.akazukin.graph.algorithm.algorithms.akima2;
+package org.akazukin.graph.algorithm.algorithms.makima;
 
 import org.akazukin.graph.algorithm.builder.IGraphBuilder;
 import org.akazukin.graph.algorithm.builder.IGraphWayedBuilder;
